Add ability to use complex data types and casts
Currently Sushi inserts rows straight into the database, which does not allow any "complex" data to be saved in rows, not even an array. Think of the use case 
[ "county" => "SomeCounty", "postCodes"=>[111,222,333]]

This change adds the ability to use $casts, which allows Eloquent to cast arrays and other data as necessary into/from a string for saving to the database.